Fixed Loading Game Score Bug

+++ src/main/java/oogasalad/model/api/data/ParserPlayer.java
@@ -14,7 +14,7 @@ public record ParserPlayer(@JsonProperty("player_id") int playerId,
                            @JsonProperty("my_scoreable") List<Integer> myScoreable,
 
                            @JsonProperty("my_controllable") List<Integer> myControllable,
-                           double score,
+                           int score,
                            @JsonProperty("active_strikeable") int activeStrikeable) {
 
 }
